AP, Invulns, FNP
Simplify the FNP code, add support for All is Dust, and handle AP and invulns using the standard 40k approach for now.

Also added a handful of new weapons and targets from preview datacards. That exposed a bug in the placement of the vertical red wound max lines on the plot.